Move generic routines from DNSClient to AbstractDNSClient
